+++ test/Dtmc/Analysis/FiniteTimeSpec.hs
@@ -108,7 +108,7 @@ asTransitionKernel ::
     TransitionMatrix state ->
     Kernel.TransitionKernel state
 asTransitionKernel matrix =
-    Kernel.transitionKernel $ \source ->
+    Kernel.fromLaws $ \source ->
         checked $
             DistributionMap.fromList
                 [ (destination, stepProbability matrix source destination)
@@ -128,7 +128,7 @@ mapInitial =
 
 simpleRandomWalk :: Kernel.TransitionKernel Integer
 simpleRandomWalk =
-    Kernel.transitionKernel $ \state ->
+    Kernel.fromLaws $ \state ->
         checked
             (DistributionMap.fromList [(state - 1, 0.5), (state + 1, 0.5)])