+++ test/Dtmc/DynamicsSpec.hs
@@ -109,7 +109,7 @@ finiteInitial =
 
 kernelChain :: Kernel.TransitionKernel (Finite 3)
 kernelChain =
-    Kernel.transitionKernel $ \source ->
+    Kernel.fromLaws $ \source ->
         checked $
             DistributionMap.fromList
                 [ (destination, stepProbability finiteChain source destination)
@@ -124,7 +124,7 @@ mapInitial =
 
 simpleRandomWalk :: Kernel.TransitionKernel Integer
 simpleRandomWalk =
-    Kernel.transitionKernel $ \state ->
+    Kernel.fromLaws $ \state ->
         checked
             (DistributionMap.fromList [(state - 1, 0.5), (state + 1, 0.5)])